Only owners of event can edit or delete it

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, ValidationPipe } from "@nestjs/common";
+import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, ValidationPipe } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Attendee } from "src/attendee/attendee.entity";
 import { Like, MoreThan, Repository } from "typeorm";
@@ -21,102 +21,102 @@ export class EventsController {
    constructor(
 
     // argument for the repository class is the entity object that will be used by the repository 
-    @InjectRepository(Event)
-    private readonly repository : Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeerepository : Repository<Event>,
+    // @InjectRepository(Event)
+    // private readonly repository : Repository<Event>,
+    // @InjectRepository(Attendee)
+    // private readonly attendeerepository : Repository<Event>,
     private readonly eventsService: EventsService
    ){}
 
 
 
-    @Get()
-    // the VLidation pipes enables population of default values if they are not provided
-    @UsePipes( new ValidationPipe({transform: true}))
-    // find all will return a list
+    // @Get()
+    // // the VLidation pipes enables population of default values if they are not provided
+    // @UsePipes( new ValidationPipe({transform: true}))
+    // // find all will return a list
     
-    async findAll(@Query() filter :ListEvents){
-         // the logger class will show when the find all method is reached
-        const events = await this.eventsService
-            .getEventsWithAttendeeCountFilteredPaginated(
-                filter, {
-                    total:true,
-                    currentPage: filter.page,
-                    limit: 2
-                });
-        return events
-    }
+    // async findAll(@Query() filter :ListEvents){
+    //      // the logger class will show when the find all method is reached
+    //     const events = await this.eventsService
+    //         .getEventsWithAttendeeCountFilteredPaginated(
+    //             filter, {
+    //                 total:true,
+    //                 currentPage: filter.page,
+    //                 limit: 2
+    //             });
+    //     return events
+    // }
 
-    @Get('/practice')
-    // practice stands for the route to be used
-    async practice (){
-        return await this.repository.find({
-            // select only these two fields in the record
-            select:['id','when'],
-            // giving the comdition to be met for each record
-            // in this case where the id of a record is more than the value guven
-            where: [{ id : MoreThan(3),
+    // @Get('/practice')
+    // // practice stands for the route to be used
+    // async practice (){
+    //     return await this.repository.find({
+    //         // select only these two fields in the record
+    //         select:['id','when'],
+    //         // giving the comdition to be met for each record
+    //         // in this case where the id of a record is more than the value guven
+    //         where: [{ id : MoreThan(3),
             
-            when: MoreThan(new Date('2021-02-12T13:00:00')) 
-        }, {
-            description: Like('%meet%')
-        }],
-        take:2,
-        order: {
-            id: 'DESC'
-        }
-        });
-    }
+    //         when: MoreThan(new Date('2021-02-12T13:00:00')) 
+    //     }, {
+    //         description: Like('%meet%')
+    //     }],
+    //     take:2,
+    //     order: {
+    //         id: 'DESC'
+    //     }
+    //     });
+    // }
     
-    @Get('practice2')
-    async practice2(){
-        // const event = await this.repository.findOne({
-        //      where : { id : 1},
-        //     // the load Eager Relations options lets you turn off the eager options in the event entity class
-        //     // when this oprion is added the attendees will not be included in the result 
-        //     // loadEagerRelations : true
+    // @Get('practice2')
+    // async practice2(){
+    //     // const event = await this.repository.findOne({
+    //     //      where : { id : 1},
+    //     //     // the load Eager Relations options lets you turn off the eager options in the event entity class
+    //     //     // when this oprion is added the attendees will not be included in the result 
+    //     //     // loadEagerRelations : true
 
-        //     // relations options takes in an array of entities that have relations with the event entity
-        //     // when this is added it will return all attendees
-        //     relations:['attendees']
-        // })
+    //     //     // relations options takes in an array of entities that have relations with the event entity
+    //     //     // when this is added it will return all attendees
+    //     //     relations:['attendees']
+    //     // })
 
-        // // when the event is not obtained, the error exception will be thrown
-        // if(!event){
-        //     throw new NotFoundException();
-        // }
+    //     // // when the event is not obtained, the error exception will be thrown
+    //     // if(!event){
+    //     //     throw new NotFoundException();
+    //     // }
 
-        // return event
+    //     // return event
 
         
-        // this will add a new attendee named william in the attendees table with an eventid matching the id of the event which is 1
-        // const event = await this.repository.findOne({
-        //     where:{id:1},
-        //     //fetches event together with relation
-        //     relations:['attendees']
-        // });
+    //     // this will add a new attendee named william in the attendees table with an eventid matching the id of the event which is 1
+    //     // const event = await this.repository.findOne({
+    //     //     where:{id:1},
+    //     //     //fetches event together with relation
+    //     //     relations:['attendees']
+    //     // });
         
-        // const attendee = new Attendee();
-        // attendee.name = 'Using Cascade ';
-        // //attendee.event = event;
+    //     // const attendee = new Attendee();
+    //     // attendee.name = 'Using Cascade ';
+    //     // //attendee.event = event;
 
-        // // pushes the attendee to the list of attendees in the events table
-        // event.attendees.push(attendee);
-        // // save using the event repository
-        // await this.repository.save(event);
+    //     // // pushes the attendee to the list of attendees in the events table
+    //     // event.attendees.push(attendee);
+    //     // // save using the event repository
+    //     // await this.repository.save(event);
 
-        // return event;
+    //     // return event;
 
-        return await this.repository.createQueryBuilder('event')
-        .select(['event.id', 'event.name','event.when'])
-        .orderBy('event.id', 'ASC')
-        .take(3)
-        .getMany()
+    //     return await this.repository.createQueryBuilder('event')
+    //     .select(['event.id', 'event.name','event.when'])
+    //     .orderBy('event.id', 'ASC')
+    //     .take(3)
+    //     .getMany()
 
-    }
+    // }
 
     @Get(':id')
-    // not including the parameter value in the @Param() function, the return value will have a kep value pair returned from the client side
+    // not including the parameter value in the @Param() function, the return value will have a key value pair returned from the client side
     // the expected return will be  "id" : "id_value" else the return will ve just the value by itself
     async findOne (@Param('id', ParseIntPipe) id: number) {
         
@@ -148,40 +148,50 @@ export class EventsController {
 
 
     @Patch(':id')
+    @UseGuards(AuthGuardJwt)
+
     async update(
         @Param('id', ParseIntPipe) id:number , 
         // this is local validation ->   new ValidationPipe({groups:['update']})
         
-        @Body() new_data: UpdateEventDto ) {
+        @Body() updates: UpdateEventDto, @CurrentUser() user : User ) {
         // the UpdateEventDto now has all variables as optional
         // all properties are being extended from the CreateEventDto 
-        new_data.address
-        new_data.description
+      
 
-        const event = await this.repository.findOne({where: { id:id }});
+        const event = await this.eventsService.getEvent(id);
 
         if(!event){
             throw new NotFoundException();
         }
-         
-        return await this.repository.save({
-            ...event,
-            ... new_data,
-            when: new_data.when? new Date(new_data.when) : event.when
 
-        });
+        if(user.id != event.organizerId){
+            throw new ForbiddenException(null, 'You are not allowed to perform this action');
+        }
+         
+        await this.eventsService.updateEvent(event, updates);
     }
 
+    
     @Delete(':id')
     @HttpCode(204)
+    // must have a valid token to delete
+    @UseGuards(AuthGuardJwt)
     async remove(
-        @Param('id', ParseIntPipe) id:number) { 
-        const result =await this.eventsService.deleteEvent(id);
-        if(result?.affected !== 1){
+        @Param('id', ParseIntPipe) id:number, 
+        @CurrentUser() user :User) { 
 
+        const event = await this.eventsService.getEvent(id);
+
+        if(!event){
             throw new NotFoundException();
-        
         }
+
+        if(user.id != event.organizerId){
+            throw new ForbiddenException(null, 'You are not allowed to perform this action');
+        }
+        
+        await this.eventsService.deleteEvent(id);
 
     }
 }
